API/getVends/requirements.txt
pandas
requests
google_cloud_storage
google_cloud_pubsub
google.api_core
functions_framework
pyarrow
fastparquet
jsonmerge